feat: add mikar provider
...and refactor deer provider to support other fleetster based providers like mikar
