+++ x2gbfs/providers/__init__.py
@@ -6,5 +6,6 @@ from .fleetster import FleetsterAPI
 from .free2move import Free2moveAPI, Free2moveProvider
 from .gbfslight import GbfsLightProvider
 from .lastenvelo_fr import LastenVeloFreiburgProvider
+from .mikar import MikarProvider
 from .moqo import MoqoProvider
 from .noi import NoiProvider
